refactor: Rename http package to httptransport and update sync references to synckit
