+++ example/client/client.go
@@ -7,7 +7,7 @@ import (
 
 	synckit "github.com/c0deZ3R0/go-sync-kit"
 	"github.com/c0deZ3R0/go-sync-kit/example/metrics"
-	"github.com/c0deZ3R0/go-sync-kit/storage/sqlite"
+"github.com/c0deZ3R0/go-sync-kit/transport/httptransport"
 	transport "github.com/c0deZ3R0/go-sync-kit/transport/http"
 )
